pyproject.toml
[project]
name = "bsc-cs-pren1"
version = "0.1.0"
description = "PREN1 @ HSLU"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scipy>=1.15.3",
    "tk>=0.1.0",
]
